requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
pixie-python==4.3.0